requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Browser Automation
playwright==1.40.0
selenium==4.16.0

# Web Scraping
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0

# LLM Integration
ollama==0.1.6
openai==1.6.1  # For Ollama API compatibility

# Email & Networking
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
dnspython==2.4.2  # For email verification

# Scheduling
APScheduler==3.10.4

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Database
sqlite3  # Built-in with Python

# Document Generation
python-docx==1.1.0
reportlab==4.0.7  # For PDF generation
PyPDF2==3.0.1

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
colorama==0.4.6
tabulate==0.9.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Logging
loguru==0.7.2
